==> Lipidomics/scripts/01_load_data.Rmd <==
---
title: "Lipidomics Baseline Analysis"
subtitle: "Load and Save Raw Data"
author: "Sofie Olund Villumsen"
date: "31/8/2021"
output: html_document
---

R markdown doc analyzing and exploring lipidomics data in combination with clinical from the type 2 diabetes overfeeding study of 48 men with mean age of 38 years.

Clear workspace
```{r}
rm(list = ls()) 
```

Load packages
```{r message=FALSE}
library(tidyverse)
library(readxl)
```

Load data
```{r message=FALSE}
path <- "L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie/Baseline_study/Lipidomics"
  
setwd(path)

lipid_data <- read_xlsx(paste(path, "data/raw_data/Kopi_af_Overfeeding_Plasma_Lipidomics.xlsx", sep="/")) %>% 
  write_csv(paste(path, "data/01_lipidomics_data.csv", sep="/"))

clinical_data <- read_xlsx(paste(path, "data/raw_data/hcot_clinical_data.xlsx", sep="/")) %>% 
  write_csv(paste(path, "data/01_clinical_data.csv", sep="/"))

lipid_names <- read_xlsx(paste(path, "data/raw_data/lipid_db_id.xlsx", sep="/")) %>% 
  write_csv(paste(path, "data/01_lipid_db_id.csv", sep="/"))
```
