==> Lipidomics/scripts/03_EDA.Rmd <==
---
title: "Lipidomics Baseline Analysis"
subtitle: "Exploratory Data Analysis (EDA)"
author: "Sofie Olund Villumsen"
date: "31/8/2021"
output: 
  html_document:
    toc: true
    toc_depth: 2
---

# Initial settings
```{r echo=FALSE}
# Clear workspace
rm(list = ls()) 
```

Load packages
```{r message=FALSE}
library(tidyverse)
library(PCAtools) 
library(broom) #K-means
```

Load data
```{r message=FALSE}
path <- "L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie/Baseline_study/Lipidomics"
setwd(path)

lipid_data <- read_csv(paste(path, "data/02_lipidomics_data_log2_zscore.csv", sep="/"))
combined_data <- read_csv(paste(path, "data/02_combined_data_log2_zscore.csv", sep="/")) 

toplipids_data <- read_tsv(paste(path, "data/04_lipidomics_top_lipids.tsv", sep="/")) 
```

Seed: Set a seed for reproducibility of data. Seed sampled with -> sample(1e6, 1)
```{r}
set.seed(422053)
```

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

# PCA
- Create PCA object
- Check for outliers

```{r echo=FALSE, message=FALSE}
#### Create PCA object
combined_data <- combined_data %>% 
  filter(Label == "A") %>% 
  unite("id_bwcat_NAFLD", c(Sample_ID, bwcat_NAFLD), remove = FALSE)  

# Prepare data format of lipidomics data 
lipid_names <- colnames(lipid_data)[3:ncol(lipid_data)]
lipid_df <- combined_data %>% 
  select(all_of(lipid_names)) %>% 
  t()
colnames(lipid_df) <- combined_data$id_bwcat_NAFLD #colnames(metabo_data)[4:ncol(metabo_data)]
rownames(lipid_df) <- lipid_names

# Prepare data format of clinical data 
clin_data <- combined_data %>%  
  filter(Label == "A") %>% 
  select(everything(), -all_of(colnames(lipid_data))) %>% 
  mutate(bwcat_NAFLD_binary = case_when(bwcat_NAFLD == 'NBW_FALSE' ~ 0,
                                        bwcat_NAFLD == 'LBW_FALSE' ~ 1,
                                        bwcat_NAFLD == 'LBW_TRUE' ~ 2),
         bwcat_binary = case_when(bwcat == 'NBW' ~ 0,
                                  bwcat == 'LBW' ~ 1),
         NAFLD_binary = case_when(NAFLD == FALSE ~ 0,
                                  NAFLD == TRUE ~ 1))  %>% 
  as.data.frame()
rownames(clin_data) <- combined_data$id_bwcat_NAFLD

clin_data$bwcat_NAFLD <- factor(clin_data$bwcat_NAFLD , levels=c("NBW_FALSE", "LBW_FALSE", "LBW_TRUE"))

pca_object <- pca(as.data.frame(lipid_df), metadata = as.data.frame(clin_data), removeVar = 0.1)
```


```{r echo=FALSE, warning=FALSE}
# Biplot
biplot_ <- biplot(pca_object,
    colby = 'bwcat_NAFLD',
    hline = 0, vline = 0,
    legendPosition = 'right', 
    title = "PCA plot")
#biplot_
#ggsave(paste(path, "results/03_01_biplot.png", sep="/"), plot = biplot_, device = "png", width = 10, height = 10)
```


```{r echo=FALSE, warning=FALSE}
biplot_loadings <- biplot(pca_object, showLoadings = TRUE,
    labSize = 3, pointSize = 3, sizeLoadingsNames = 3, lab = NULL, colby = 'bwcat_NAFLD', 
    title = "Bi-plot with loadings",
    # ellipse config
      ellipse = TRUE,
      ellipseType = 't',
      ellipseLevel = 0.95,
      ellipseFill = FALSE,
      ellipseAlpha = 1/4,
      ellipseLineSize = 1.0,
    legendPosition = 'right',
    xlim = c(-70,55), ylim = c(-25,20))
biplot_loadings
ggsave(paste(path, "results/03_02_biplot_loadings.png", sep="/"), plot = biplot_loadings, device = "png", width = 15, height = 10)
```

**Bi-plot:** A plot that plots both variables and observations (samples) in the same space. The variables are indicated by arrows drawn from the origin, which indicate their ‘weight’ in different directions. No outliers are detected.

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

# PCA - Outlier removed
- Create PCA object
- Scree plot - check the variance of data
- Loading plots - identify how the PCs are loaded

Remove outlier and create new PCA plot (outlier = "050_NBW_FALSE" from baseline)
```{r echo=FALSE, message=FALSE}
combined_data_baseline_outlier <- combined_data %>% 
  filter(Label == "A") %>%
  unite("id_bwcat_NAFLD", c(Sample_ID, bwcat_NAFLD), remove = FALSE) %>% 
  filter(id_bwcat_NAFLD != "008_NBW_FALSE") # One outlier has been removed

# Prepare data format of metabolomic data 
lipid_names <- colnames(lipid_data)[3:ncol(lipid_data)]
lipid_df <- combined_data_baseline_outlier %>% 
  select(all_of(lipid_names)) %>% 
  t()
colnames(lipid_df) <- combined_data_baseline_outlier$id_bwcat_NAFLD #colnames(metabo_data)[4:ncol(metabo_data)]
rownames(lipid_df) <- lipid_names

# Prepare data format of clinical data 
clin_data <- combined_data_baseline_outlier %>%  
  select(everything(), -all_of(colnames(lipid_data))) %>% 
  mutate(bwcat_NAFLD_binary = case_when(bwcat_NAFLD == 'NBW_FALSE' ~ 0,
                                        bwcat_NAFLD == 'LBW_FALSE' ~ 1,
                                        bwcat_NAFLD == 'LBW_TRUE' ~ 2),
         bwcat_binary = case_when(bwcat == 'NBW' ~ 0,
                                  bwcat == 'LBW' ~ 1),
         NAFLD_binary = case_when(NAFLD == FALSE ~ 0,
                                  NAFLD == TRUE ~ 1))  %>% 
  as.data.frame()
rownames(clin_data) <- combined_data_baseline_outlier$id_bwcat_NAFLD

clin_data$bwcat_NAFLD <- factor(clin_data$bwcat_NAFLD , levels=c("NBW_FALSE", "LBW_FALSE", "LBW_TRUE"))

pca_object <- pca(as.data.frame(lipid_df), metadata = as.data.frame(clin_data), removeVar = 0.1)
```


```{r echo=FALSE, warning=FALSE}
#### Biplot
#- Different kinds of bi-plots
#- Pairs plot - explore potentially informative PCs
#- Loading plots - identify how the PCs are loaded

biplot_ <- biplot(pca_object,
    colby = 'bwcat_NAFLD',
    hline = 0, vline = 0,
    legendPosition = 'right', 
    title = "PCA plot - outier removed")
#biplot_
##ggsave(paste(path, "results/03_03_biplot_outlierremoved.png", sep="/"), plot = biplot_, device = "png", width = 10, height = 10)
```

```{r echo=FALSE, warning=FALSE}
biplot_loadings <- biplot(pca_object, showLoadings = TRUE,
    labSize = 3, pointSize = 3, sizeLoadingsNames = 3, lab = NULL, colby = 'bwcat_NAFLD', 
    title = "PCA plot - outier removed",
    # ellipse config
      ellipse = TRUE,
      ellipseType = 't',
      ellipseLevel = 0.95,
      ellipseFill = FALSE,
      ellipseAlpha = 1/4,
      ellipseLineSize = 1.0,
    legendPosition = 'right',
    xlim = c(-70,55), ylim = c(-25,20))
biplot_loadings
ggsave(paste(path, "results/03_04_biplot_loadings_outlierremoved.png", sep="/"), plot = biplot_loadings, device = "png", width = 15, height = 10)
```

**Bi-plot:** A plot that plots both variables and observations (samples) in the same space. The variables are indicated by arrows drawn from the origin, which indicate their ‘weight’ in different directions. 

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

## Scree plot: PCs explaining the variance
Check how many PC covers 90% of the variance
```{r echo=FALSE}
which(cumsum(pca_object$variance) > 90)[1] 
```


```{r warning=FALSE, echo=FALSE}
horn <- parallelPCA(as.data.frame(lipid_df))
#elbow <- findElbowPoint(pca_object$variance)

screeplot_ <- screeplot(pca_object, axisLabSize = 13, titleLabSize = 22, 
                        components = getComponents(pca_object, 1:18),
                        vline = c(horn$n)) +
  geom_label(aes(x = horn$n + 1, y = 50, label = 'Horn\'s', vjust = 1, hjust = 1, size = 8)) 
screeplot_
##ggsave(paste(path, "results/03_05_screeplot.png", sep="/"), plot = screeplot_, device = "png", width = 10, height = 10)
```

**Scree plot:** A plot including PCs showing the accumulative proportion of explained variation covering 90%. To determine the optimum number of PCs and to retain PCs Horn's parallel analysis is used.

**Comments:** based on the outcome an optimal number of PCs to proceed with is either 6 or 8. PC1 only accounts for 17% of the maximum variance in the metabolite data set, which is not much. 

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------


```{r echo=FALSE}
#### Pairs plot: explores potentially informative PCs
pairsplot_ <- pairsplot(pca_object,
    components = getComponents(pca_object, c(1,2,3,4,5,6,7)),
    triangle = FALSE,
    hline = 0, vline = 0,
    pointSize = 0.8,
    gridlines.major = FALSE, gridlines.minor = FALSE,
    colby = 'bwcat_NAFLD',
    axisLabSize = 14,
    title = 'Pairs plot', 
    plotaxes = TRUE,
    margingaps = unit(c(-0.01, -0.01, -0.01, -0.01), 'cm'))
#pairsplot_
#ggsave(paste(path, "results/03_06_pairsplot.png", sep="/"), plot = pairsplot_, device = "png", width = 20, height = 10)
```



## Loadings plot

Determine variables that drive variation among each PC

- Check how this loadings plot corresponds to the biplot loadings - they should match up for the top hits (Tryptophan, Oleic acid)
- If, on the bi-plot or pairs plot, there is evidence that 1 or more PCs are segregating a factor of interest, we can explore further the metabolites that are driving these differences along each PC.
- For each PC of interest, ‘plotloadings’ determines the variables falling within the top or bottom 5% of the loadings range, and then creates a final consensus list of these. These variables are then plotted.
```{r echo=FALSE, warning=FALSE}
loadingsplot <- plotloadings(pca_object, components = getComponents(pca_object, 1:horn$n), 
                             rangeRetain = 0.01,
                             labSize = 4.0,
                             title = 'Loadings plot from PCA',
                             caption = 'Top 1% variables',
                             shape = 24,
                             col = c('limegreen', 'black', 'red3'),
                             drawConnectors = TRUE)
loadingsplot
ggsave(paste(path, "results/03_xx_loadings.png", sep="/"), plot = loadingsplot, device = "png", width = 15, height = 15)
```

**Comments:** there is not any eye catching or especially interesting findings. However, some of the metabolites are more responsible for variation along different PCs (e.g. Linoleic acid and Oleic acid along PC1). The previous bi-plots did not really seperate the groups from each other, which correponds to this result. 


```{r echo=FALSE, message=FALSE}
#### **Extra:** plotting component loadings 
# Plot of absolute values and modified point sizes to be proportional to the loadings. 
p_loadings <- plotloadings(pca_object,
  components = getComponents(pca_object, 1:horn$n),
  rangeRetain = 0.01,
  labSize = 4.0,
  absolute = FALSE,
  title = 'Loadings plot',
  subtitle = 'Misc PCs',
  caption = 'Top 1% variables',
  shape = 23, shapeSizeRange = c(1, 16),
  col = c('white', 'pink'),
  drawConnectors = FALSE)
```



---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------


# PCs vs. clinical data
- Correlation of the PCs back to the clinical data
- Further exploration of the PCs can come through correlations with clinical data. This is also a mostly untapped resource in the era of ‘big data’ and can help to guide an analysis down a particular path.
- We now proceed with PCs that account for the "horn" cutoff and then explore further the PCs that have statistically significant correlations.


## Eigencor plot
Correlation between PC[1-6] and the clinical variables. They have been correted for multiple testing by the BH method.
```{r echo=FALSE, message=FALSE}
eigencorplot(pca_object, components = getComponents(pca_object, 1:horn$n),
    metavars = c("bwcat_binary", "bwcat_NAFLD_binary", "NAFLD_binary",
      "height_day1","weight_day1","bmi_day1","navelcirc_day1","iliaccirc_day1","hipcirc_day1","sysbp_day1","diabp_day1",
      "pulsebp_day1","WHR_day1","age_day1","map_day1","total_BMD","vat_mass","vat_volume",
      "AG_fatratio","android_tissuefatpercentage","gynoid_tissuefatpercentage",
      "fatmass_total","leanmass_total","total_tissuefatpercetage","liverfatpercent","VO2_20sec_max","VCO2_20sec_max", 
      "RQ_20sec_max","fitness_lvl","bas_glu", "bas_ins","bas_cpe","a1c_day2", "cholesterol_day2","hdl_day2","ldl_day2",
      "triglyc_day2","alat_day2","asat_day2","ggt_day2","factiiviix_day2","homa_ir",
      "adiponectin","leptin","ghrelin_total","ghrelin_active","FGF.21","GLP.1_total",
      "GLP.1_active","GLP.1_inactive","GIP_total","GIP_active","GIP_inactive","pyy",
      "glucagon","proinsulin","FGF.23","c.peptide","PP","LH","BAFF","FSH","BDNF","beta.NGF"),
    cexCorval = 0.7,
    colCorval = 'black',
    fontCorval = 1.5,
    posLab = 'bottomleft',
    rotLabX = 45,
    posColKey = 'top',
    cexLabColKey = 1,
    scale = TRUE,
    colFrame = 'white',
    plotRsquared = FALSE,
    corMultipleTestCorrection = "BH",
    main = 'PC1-6 clinical correlations')

png(paste(path, "results/03_08_eigencorplot_clinical_PC.png", sep="/"),
    width = 6*200,        # 5 x 300 pixels
    height = 6*450,
    res = 200,            # 300 pixels per inch
    pointsize = 150)      # font size)   
eigencorplot(pca_object, components = getComponents(pca_object, 1:horn$n),
    metavars = c("bwcat_binary", "bwcat_NAFLD_binary", "NAFLD_binary",
      "height_day1","weight_day1","bmi_day1","navelcirc_day1","iliaccirc_day1","hipcirc_day1","sysbp_day1","diabp_day1",
      "pulsebp_day1","WHR_day1","age_day1","map_day1","total_BMD","vat_mass","vat_volume",
      "AG_fatratio","android_tissuefatpercentage","gynoid_tissuefatpercentage",
      "fatmass_total","leanmass_total","total_tissuefatpercetage","liverfatpercent","VO2_20sec_max","VCO2_20sec_max", 
      "RQ_20sec_max","fitness_lvl","bas_glu", "bas_ins","bas_cpe","a1c_day2", "cholesterol_day2","hdl_day2","ldl_day2",
      "triglyc_day2","alat_day2","asat_day2","ggt_day2","factiiviix_day2","homa_ir",
      "adiponectin","leptin","ghrelin_total","ghrelin_active","FGF.21","GLP.1_total",
      "GLP.1_active","GLP.1_inactive","GIP_total","GIP_active","GIP_inactive","pyy",
      "glucagon","proinsulin","FGF.23","c.peptide","PP","LH","BAFF","FSH","BDNF","beta.NGF"),
    cexCorval = 0.7,
    colCorval = 'black',
    fontCorval = 1.5,
    posLab = 'bottomleft',
    rotLabX = 45,
    posColKey = 'top',
    cexLabColKey = 1,
    scale = TRUE,
    colFrame = 'white',
    plotRsquared = FALSE,
    corMultipleTestCorrection = "BH",
    main = 'PC1-6 clinical correlations')
dev.off()
```

**Comments:** after multiple correction with BH (FDR), we are left with one negative correlations (r > -0.70), this relationship is however not significantly associated with each other (PC1 vs. GIP_inactive). Furthermore, we observe that we have two moderatly positive correlations (r > 0.52), which are also significantly associated with each other (FDR < 0.05) (PC6 vs. cholesterol_day2 and PC6 vs. triglyceride_day2). This indicates that PC1 and PC6 are the most interesting PCs, with PC6 having statistical significant association and correlation (FDR < 0.05, r > 0.52) to cholesterol and triglycerides. When looking at PC6 in the loadings plot we see that the metabolite driving most of the variance between the groups is the metabolite cholesterol.


```{r echo=FALSE, message=FALSE}
#### **Eigencor plot 2**
#Different cut-offs for statistical significance exists, such as p-value adjustment, plot R-squared values, and specify correlation method:

png(paste(path, "results/03_09_eigencorplot_clinical_PC_2.png", sep="/"),
    width = 6*300,        # 5 x 300 pixels
    height = 6*300,
    res = 200,            # 300 pixels per inch
    pointsize = 150)      # font size)   
eigencorplot(pca_object,
    components = getComponents(pca_object, 1:horn$n),
    metavars = c("bwcat_binary", "bwcat_NAFLD_binary", "NAFLD_binary",
      "height_day1","weight_day1","bmi_day1","navelcirc_day1","iliaccirc_day1","hipcirc_day1","sysbp_day1","diabp_day1",
      "pulsebp_day1","WHR_day1","age_day1","map_day1","total_BMD","vat_mass","vat_volume",
      "AG_fatratio","android_tissuefatpercentage","gynoid_tissuefatpercentage",
      "fatmass_total","leanmass_total","total_tissuefatpercetage","liverfatpercent","VO2_20sec_max","VCO2_20sec_max", 
      "RQ_20sec_max","fitness_lvl","bas_glu", "bas_ins","bas_cpe","a1c_day2", "cholesterol_day2","hdl_day2","ldl_day2",
      "triglyc_day2","alat_day2","asat_day2","ggt_day2","factiiviix_day2","homa_ir",
      "adiponectin","leptin","ghrelin_total","ghrelin_active","FGF.21","GLP.1_total",
      "GLP.1_active","GLP.1_inactive","GIP_total","GIP_active","GIP_inactive","pyy",
      "glucagon","proinsulin","FGF.23","c.peptide","PP","LH","BAFF","FSH","BDNF","beta.NGF"),
    col = c('white', 'cornsilk1', 'gold', 'forestgreen', 'darkgreen'),
    cexCorval = 0.7,
    fontCorval = 1.5,
   # posLab = 'all',
    rotLabX = 45,
    scale = TRUE,
    main = bquote(Principal ~ component ~ Pearson ~ r^2 ~ clinical ~ correlates),
    plotRsquared = TRUE,
    corFUN = 'pearson',
    corUSE = 'pairwise.complete.obs',
    corMultipleTestCorrection = 'BH',
    signifSymbols = c('****', '***', '**', '*', ''),
    signifCutpoints = c(0, 0.0001, 0.001, 0.01, 0.05, 1))
dev.off()
```

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

## Loadings for PC1
Plot the component loadings for PC1
```{r echo=FALSE}
PC1_plotloadings <- plotloadings(pca_object,
  components = getComponents(pca_object, c(1)),
  rangeRetain = 0.1, absolute = FALSE,
  col = c('black', 'pink', 'red4'),
  drawConnectors = TRUE, labSize = 4, title = "Loadings for PC1") + coord_flip()
PC1_plotloadings
ggsave(paste(path, "results/03_10_PC1_plotloadings.png", sep="/"), plot = PC1_plotloadings, device = "png", width = 15, height = 15)

```

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------


## Loadings for PC4
Plot the component loadings for PC6
```{r echo=FALSE}
PC4_plotloadings <- plotloadings(pca_object,
  components = getComponents(pca_object, c(4)),
  rangeRetain = 0.1, absolute = FALSE,
  col = c('black', 'pink', 'red4'),
  drawConnectors = TRUE, labSize = 4, title = "Loadings for PC4") + coord_flip()
PC4_plotloadings
ggsave(paste(path, "results/03_11_PC4_plotloadings.png", sep="/"), plot = PC4_plotloadings, device = "png", width = 15, height = 15)

```

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

# PCA with Broom package
- Create pca object for dimensionality reduction
- PCA plot, investigating different distributions depending on different variables

```{r message=FALSE, echo=FALSE}
#### Create pca object for dimensionality reduction

# Get a list of metabolites
lipid_names <- colnames(lipid_data)[3:ncol(lipid_data)]

# Create PCA object on log2 transformed and scaled (z-score/auto-scale) variables
data_pca <- combined_data_baseline_outlier %>% 
  select(-everything(), all_of(lipid_names)) %>% 
  prcomp(scale. = TRUE)

# Tidy data in order to get proper data table format
data_pca_tidy <- data_pca %>% 
  tidy("pcs")
```

PCA plot of the principal components base on variable: bwcat_NAFLD - PC1 and PC6 is used as they were most informative (seen in previous analyses)

```{r echo=FALSE}
# Augment data in order to get a complete table with original values and PC values. 
data_pca_aug <- data_pca %>% 
  augment(combined_data_baseline_outlier)

# Adding percentage to the PCA plot
x <- data_pca_tidy %>% 
  filter(PC == 1) %>% 
  pull(percent)
x <- str_c("PC1 (", round(x*100, 2), "%)")

y <- data_pca_tidy %>% 
  filter(PC == 4) %>% 
  pull(percent)
y <- str_c("PC4 (", round(y*100, 2), "%)")

#data_pca_aug$bwcat_NAFLD <- factor(data_pca_aug$bwcat_NAFLD , levels=c("NBW_FALSE", "LBW_FALSE", "LBW_TRUE"))

# Plot PCA with medical condition as labels
pca_bwcat_NAFLD <- data_pca_aug %>% 
  ggplot(aes(x = .fittedPC1,
             y = .fittedPC4,
             colour = bwcat_NAFLD)) + 
  geom_point(size = 3, alpha = 0.5) + 
  labs(x = x, y = y, title = "PCA on LBW and NBW with and without NAFLD", color = "Condition") +
  #scale_color_manual(values = c("#32CD32", "#6495ED")) + 
  stat_ellipse(level = 0.95) # illustrating the 95% CI interval
pca_bwcat_NAFLD
ggsave(paste(path, "results/03_12_pca_PC1vsPC6.png", sep="/"), plot = pca_bwcat_NAFLD, device = "png", width = 6.17, height = 3.1)
```

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

---------------------------------------------------------------------------------------------------------------------------------

# K-means: Cluster Analysis
- Use principal components (from PCA) for a cluster analysis

```{r echo=FALSE, message=FALSE}
# Perform kmeans
data_kmeans <- data_pca_aug %>%
  select(contains("PC")) %>% 
  kmeans(centers = 3, nstart = 24)

# Add cluster column to augmented pca 
data_kmeans_aug <- data_kmeans %>%
  augment(data_pca_aug) %>%
  rename(Cluster = .cluster)

# Plot kmeans on two first principal components
kmeans_condition <- data_kmeans_aug %>% 
  ggplot(aes(x = .fittedPC1,
             y = .fittedPC4,
             colour = Cluster)) +
  geom_point(size = 4, alpha = 0.5) +
  labs(x = x, y = y, title = "K-means") +
  stat_ellipse(level = 0.95) 
kmeans_condition
ggsave(paste(path, "results/03_13_kmeans.png", sep="/"), plot = kmeans_condition, device = "png", width = 6.17, height = 3.1)

```

**Comments:** the best cluster seperation is obtained when using 2 clusters. This might be due to the fact that LBW and NBW without NAFLD show same values compared to LBW with NAFLD. 

------------------------------------------------------------------------------------------------------------------------------

------------------------------------------------------------------------------------------------------------------------------

------------------------------------------------------------------------------------------------------------------------------
